+++ wallet/wallet.h
@@ -3,6 +3,7 @@
 
 #include "config.h"
 #include "db.h"
+#include <bitcoin/tx.h>
 #include <ccan/crypto/shachain/shachain.h>
 #include <ccan/list/list.h>
 #include <ccan/tal/tal.h>
@@ -210,5 +211,11 @@ bool wallet_peer_by_nodeid(struct wallet *w, const struct pubkey *nodeid,
  * loaded from the database to the list without checking.
  */
 bool wallet_channels_load_active(struct wallet *w, struct list_head *peers);
+
+/**
+ * wallet_extract_owned_outputs - given a tx, extract all of our outputs
+ */
+int wallet_extract_owned_outputs(struct wallet *w, const struct bitcoin_tx *tx,
+				 u64 *total_satoshi);
 
 #endif /* WALLET_WALLET_H */
